+++ Borrow_Book_Requirement/ak_library_management/models/borrow_transaction_history.py
@@ -1,7 +1,8 @@
 # -*- coding: utf-8 -*-
 
-from odoo import models, fields,api
 from datetime import timedelta
+
+from odoo import models, fields, api
 from odoo.models import ValidationError
 
 
@@ -11,18 +12,27 @@ class BorrowTransactionHistory(models.Model):
     This model maintains detailed records of book borrowing activities including
     customer information, borrowed books, transaction dates, and deposit details.
     """
-    _name = 'borrow.transaction.history'
-    _description = 'Borrow Transaction History'
-    _rec_name = 'customer_id'
 
-    customer_id = fields.Many2one(comodel_name='res.partner', string='Customer', required=True)
-    books_ids = fields.Many2many(comodel_name='product.template', string='Books', domain=[('is_library_book', '=', 'True')])
-    borrow_start_date = fields.Date(string='Borrow Start Date', default=fields.datetime.now())
-    borrow_end_date = fields.Date(string='Borrow End Date', required=True)
-    deposit_amount = fields.Float(string='Deposit Amount')
-    is_member = fields.Boolean(related='customer_id.is_member')
+    _name = "borrow.transaction.history"
+    _description = "Borrow Transaction History"
+    _rec_name = "customer_id"
 
-    @api.constrains('borrow_start_date', 'borrow_end_date')
+    customer_id = fields.Many2one(
+        comodel_name="res.partner", string="Customer", required=True
+    )
+    books_ids = fields.Many2many(
+        comodel_name="product.template",
+        string="Books",
+        domain=[("is_library_book", "=", "True")],
+    )
+    borrow_start_date = fields.Date(
+        string="Borrow Start Date", default=fields.datetime.now()
+    )
+    borrow_end_date = fields.Date(string="Borrow End Date", required=True)
+    deposit_amount = fields.Float(string="Deposit Amount")
+    is_member = fields.Boolean(related="customer_id.is_member")
+
+    @api.constrains("borrow_start_date", "borrow_end_date")
     def validate_borrow_dates(self):
         """
         Validates that the borrow end date is after the start date.
@@ -32,9 +42,9 @@ class BorrowTransactionHistory(models.Model):
             ValidationError: When borrow end date is earlier than start date
         """
         if self.filtered(lambda r: r.borrow_start_date > r.borrow_end_date):
-            raise ValidationError('Borrow end date not less then the start date')
+            raise ValidationError("Borrow end date not less then the start date")
         if self.filtered(lambda r: r.deposit_amount <= 0 and not r.is_member):
-            raise ValidationError('Deposit Amount must be greater than zero')
+            raise ValidationError("Deposit Amount must be greater than zero")
 
     def _get_wizard_popup(self, title, message):
         """
@@ -44,14 +54,12 @@ class BorrowTransactionHistory(models.Model):
         return type: dict
         """
         return {
-            'name': title,
-            'type': 'ir.actions.act_window',
-            'res_model': 'borrow.books.wizard',
-            'view_mode': 'form',
-            'target': 'new',
-            'context': {
-                'default_message': message
-            },
+            "name": title,
+            "type": "ir.actions.act_window",
+            "res_model": "borrow.books.wizard",
+            "view_mode": "form",
+            "target": "new",
+            "context": {"default_message": message},
         }
 
     def action_borrow_books(self):
@@ -64,68 +72,73 @@ class BorrowTransactionHistory(models.Model):
         # Condition: Customer is not trustworthy
         if self.customer_id.not_trust_worthy:
             return self._get_wizard_popup(
-                title='Customer Not Trustworthy',
-                message="The customer is marked as not trustworthy. Are you sure you want to continue?"
+                title="Customer Not Trustworthy",
+                message="The customer is marked as not trustworthy. Are you sure you want to continue?",
             )
 
+
         # Condition: Books are out of stock
-        out_of_stock_books = self.books_ids.filtered(lambda book: book.qty_available == 0)
+        out_of_stock_books = self.books_ids.filtered(
+            lambda book: book.qty_available == 0
+        )
         if out_of_stock_books:
-            return self._get_wizard_popup(
-                title='Book Out of Stock',
-                message=f"The book '{out_of_stock_books[0].name}' is out of stock. Do you want to proceed?"
-            )
+            book_names = ", ".join(out_of_stock_books.mapped("name"))
+            message = f"The following books are out of stock: {book_names}. Do you want to proceed?"
+            return self._get_wizard_popup(title="Book Out of Stock", message=message)
+
+        # decrease a stock of product when borrowed
+        for rec in self.books_ids.filtered(lambda book: book.qty_available):
+            loc = self.env['stock.quant'].search([('product_tmpl_id.id', '=', rec.id)], limit=1)
+            self.env['stock.quant']._update_available_quantity(loc.product_id, loc.location_id,
+                                                                    quantity=-1)
+
 
         # If the customer is trying to borrow 5 or more books
         if len(self.books_ids) >= 5:
-            search_recd = self.search([('customer_id.id', "=", self.customer_id.id)], order='id desc', offset=1)
+            books_record = self.search(
+                [("customer_id.id", "=", self.customer_id.id)],
+                order="id desc",
+                offset=1,
+            )
             books_name = []
-            [books_name.append(book.name) for rec in search_recd
-             for book in rec.books_ids if book.name not in books_name]
+            books_name = books_record.mapped('books_ids').filtered(lambda b: b.name not in books_name).mapped(
+                'name')
 
             if books_name:
-                return self._get_wizard_popup(title='warning',
-                            message=f"Customer already has [{len(search_recd)}] open borrow transactions "
-                           f"with {books_name} books. "
-                           f"Are you sure you want to borrow more books?")
-
-            return self._get_wizard_popup(title='warning',
-                        message="Are you sure you want to allow "
-                       "borrowing more than 5 books for this customer?")
-
-        # decrease a stock of product when borrowed
-        for rec in self.books_ids:
-            if rec.qty_available:
-                product_id = self.env['product.product'].search([('name', '=', rec.name),('default_code', '=', rec.default_code)])
-                loc = self.env['stock.quant'].search([
-                    ('product_id', '=', product_id.id),
-                    ('location_id.usage', '=', 'internal')
-                ], limit=1)
-                self.env['stock.quant']._update_available_quantity(
-                    product_id,
-                    loc.location_id,
-                    quantity=-1
+                return self._get_wizard_popup(
+                    title="warning",
+                    message=f"Customer already has [{len(books_record)}] open borrow transactions "
+                            f"with {books_name} books. "
+                            f"Are you sure you want to borrow more books?",
                 )
+
+            return self._get_wizard_popup(
+                title="warning",
+                message="Are you sure you want to allow "
+                        "borrowing more than 5 books for this customer?",
+            )
+
 
     def notify_due_returns(self):
-            """
-            This method checks for records with a borrow_end_date that is 2 days from today
-            and sends notifications to the respective customers.
-            param : self
-            return: None
-            """
-            borrowed_records = self.search([]).filtered(
-                lambda r: any(book.status == 'borrowed' for book in r.books_ids)
-                          and r.borrow_end_date == r.borrow_start_date + timedelta(days=2))
-            for record in borrowed_records:
-                self.env['bus.bus']._sendone(
-                    record.customer_id,
-                    'simple_notification',
-                    {
-                        'type': 'warning',
-                        'message': f"reminder: your book return date is {record.borrow_end_date}",
-                    }
-                )
+        """
+        This method checks for records with a borrow_end_date that is 2 days from today
+        and sends notifications to the respective customers.
+        param : self
+        return: None
+        """
+        borrowed_records = self.search([]).filtered(
+            lambda r: any(book.status == "borrowed" for book in r.books_ids)
+                      and r.borrow_end_date == r.borrow_start_date + timedelta(days=2)
+        )
+        for record in borrowed_records:
+            self.env["bus.bus"]._sendone(
+                record.customer_id,
+                "simple_notification",
+                {
+                    "type": "warning",
+                    "message": f"reminder: your book return date is {record.borrow_end_date}",
+                },
+            )
 
     def action_return_books(self):
         """
@@ -136,15 +149,20 @@ class BorrowTransactionHistory(models.Model):
         param: self
         Returns: None
         """
-        for rec in self.books_ids:
-            if rec.status == 'borrowed':
-                rec.status = 'return'
-                self.env['bus.bus']._sendone(self.customer_id, 'simple_notification', {
-                    'type': 'success',
-                    'message': f"Book '{rec.name}' has been returned successfully.",
-                })
+        borrowed_books = self.books_ids.filtered(lambda r: r.status == "borrowed")
+        borrowed_books.write({'status': 'return'})
 
-    @api.constrains('customer_id', 'books_ids')
+        for book in borrowed_books:
+            self.env["bus.bus"]._sendone(
+                self.customer_id,
+                "simple_notification",
+                {
+                    "type": "success",
+                    "message": f"Book '{book.name}' has been returned successfully.",
+                },
+            )
+
+    @api.constrains("customer_id", "books_ids")
     def _check_overdue_books(self):
         """
         Prevents customers from borrowing new books if they have overdue books.
@@ -152,18 +170,19 @@ class BorrowTransactionHistory(models.Model):
         param : self
         Returns: None
         """
-        for record in self:
-            if record.customer_id and record.books_ids:
-                overdue_count = self.env['borrow.transaction.history'].search_count([
-                    ('customer_id', '=', record.customer_id.id),
-                    ('borrow_end_date', '<', fields.Date.today()),
-                    ('books_ids.status', '=', 'borrowed'),
-                ])
+        record = self.filtered(lambda b: b.customer_id and b.books_ids)
+        overdue_count = self.env["borrow.transaction.history"].search_count(
+            [
+                ("customer_id", "=", record.customer_id.id),
+                ("borrow_end_date", "<", fields.Date.today()),
+                ("books_ids.status", "=", "borrowed"),
+            ]
+        )
 
-                if overdue_count:
-                    raise ValidationError(
-                        f"Customers with overdue books cannot borrow new ones until they return the overdue items."
-                    )
+        if overdue_count:
+            raise ValidationError(
+                f"Customers with overdue books cannot borrow new ones until they return the overdue items."
+            )
 
     def send_overdue_notices(self):
         """
@@ -172,13 +191,13 @@ class BorrowTransactionHistory(models.Model):
         param : self
         return: None
         """
-        overdue_records = self.search([
-            ('borrow_end_date', '<', fields.Date.today()),
-            ('books_ids.status', '=', 'borrowed')
-        ])
+        overdue_records = self.search(
+            [
+                ("borrow_end_date", "<", fields.Date.today()),
+                ("books_ids.status", "=", "borrowed"),
+            ]
+        )
 
-        template = self.env.ref('ak_library_management.email_template_library_overdue')
+        template = self.env.ref("ak_library_management.email_template_library_overdue")
         for record in overdue_records:
-            template.send_mail(
-                record.id,
-                force_send=True)
+            template.send_mail(record.id, force_send=True)