+++ Borrow_Book_Requirement/ak_library_management/models/library_library.py
@@ -20,6 +20,7 @@ class Library(models.Model):
         domain=[('is_library_book', '=', True)]
     )
     location = fields.Char(string='Location',tracking=True)
+    date = fields.Date(string='Date', default=fields.datetime.now())
     capacity = fields.Integer(string='Capacity')
     notes = fields.Text(string='Notes')
     book_count = fields.Integer(compute='compute_book_count')