+++ constrains_mail_thread_assignment/ak_library_management/changelog.md
@@ -4,11 +4,19 @@
 
 
 ### changes
-- add changelog in this repo 
+- add changelog 
 - update module versioning in manifest file
 - add .gitignore file in project folder
 - update a validation message in product.template model when the book is borrowed and try to click on borrowed button 
 - add chatter in library_library model view
 - remove unecessary import in product.template file
 - update a function name of revert_changes to action_revert_changes	
+
+
+## [18.0.1.0.1] - 2025-03-11
+
+### changes
+- add changelog file in module folder 
+- utilize Many2many field to store data and also retrive a data 
+- update doc string of action_revert_change function in Bulk book upload.py